Added Player class, prepared Cue to load a model
Added a player class. initialized with a player number 1 or 2, each
player has their own cue (different coloured textures), and a ball type
(see Ball.h for BALL_TYPE_SOLID/STRIPE). Initializing player will load
and store an instance of Cue.
Cue.cpp was modified to take in a player number to specify what texture
is to be used.
Added to ball.h is ball type constants for striped and solids